PEAR-2237 - move cohort bar just to analysis page
